Refactor step execution in TinkerGraphQueryExecutor
Removed the `PathWrapper` class as it is no longer needed.
Introduced the `IStepExecutor` interface to define a modular
contract for step execution. Added a static `_StepExecutors`
dictionary in `TinkerGraphQueryExecutor` for dynamic discovery
and registration of `IStepExecutor` implementations.

These changes improve extensibility and modularity in the
execution engine by enabling dynamic step execution logic.